+++ libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: tfaure <tfaure@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/11/05 16:05:35 by tfaure            #+#    #+#             */
-/*   Updated: 2017/06/21 10:44:58 by tfaure           ###   ########.fr       */
+/*   Updated: 2017/08/24 21:39:57 by tfaure           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -100,7 +100,7 @@ void				*semalloc(size_t size);
 void				err_found(char *str);
 void				free_splited_str(char **str);
 int					len_of_tab(char **tab);
-double				ft_atof(const char *str);
+float				ft_atof(const char *str);
 int					fl_to_int(float fl);
 
 #endif
